microservices/backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
mysql-connector==2.2.9
gunicorn==20.1.0
